messagebox: Redesign focus rectangle and unread marker.
Steps 5 and 6 of #22059.

This change replaces the box shadow focus rectangle with a more
subtle light blue rectangle, also also thins the unread message
indicator by 1px to only 2px.

The Figma designs show the unread indicator as 1px, but I left it
at 2px to be noticeable beside the stream color bar.

Note that this change adds `3px` of height to the messagebox to
make space for the new rectangle. More discussion on that
[here](https://chat.zulip.org/#narrow/stream/101-design/topic/redesign.3A.20.20focus.20rectangle/near/1491642).