css: Fix blue box padding.
This is intended to make space between paragraphs in a messages
and the space between single line paragraphs between two different
messages the same.

Increase margin between two `p` in a msg by 1px. This was need to
get the same space between paragraphs in the
same messages vs different messages.